refactor: table-groups 패키지 분리
